Create a layer with an arrow or segment from and to specified locations

A wrapper for ggplot2::geom_segment, ggplot2::geom_curve(), ggarchery::geom_arrowsegment(), ggforce::geom_diagonal0(), ggforce::geom_link() that makes it easier to specify the start and end of the segment using any object supported by the sfext::as_xy() function.

Usage

layer_arrow(
  mapping = NULL,
  data = NULL,
  crs = NULL,
  from,
  to,
  geom = "segment",
  ...
)

Arguments

mapping

aesthetic mapping overwritten with x, y, xend, and yend values based on provided from and to parameters.

data

Required if from or to are character vectors to sfext::as_xy()

crs

A character or numeric reference to a coordinate reference system supported by sf::st_crs() or another sf, sfc, or bbox object that is used to provide crs.

from, to

Required. Passed to x parameter of sfext::as_xy() (using nm = c("xend", "yend")) for the to parameter.

geom

Character string for geom to use c("segment", "curve", "arrowsegment", "diagonal0", "link") or a geom function.

...

Additional parameters passed to sf_bbox_point(), as_points(), or as_sfc().

See also