Map icons

An index of map icons from seven sources:

mapbox/maki

ideditor/temaki

manifestinteractive/weather-underground-icons

openstreetmap/map-icons

openstreetmap/lane-icons

Esri/calcite-point-symbols

NPMap Symbol Library

Most of these icon sources use open licenses. Maki, Temaki, and the OSM lane icons all use a CC0 license. The Weather Underground Icons use an MIT license. The OSM map icons use an unspecified PD style license. The Calcite icons are available under the Esri Master License Agreement (MLA). The NPMap Symbol Library is created by the National Park Service so is assumed to be a Public Domain work in the U.S..

This index was last updated on 2024-04-08 and may not include any icons that have been added since that date.

Usage

map_icons

Format

A tibble with 5,036 rows and 5 variables:

name

Icon name

url

Icon URL on GitHub repo

style

Icon style (blank if not applicable)

size

Icon width/height (pixels)

repo

GitHub repository for icon collection

Details

The name column is not unique so a px or source may be required when using layer_icon().