Add labels to a ggplot2 plot or map

A helper function that converts strings to glue strings for the title, subtitle, and caption. In progress.

Usage

labs_ext(
  title = ggplot2::waiver(),
  subtitle = ggplot2::waiver(),
  caption = ggplot2::waiver(),
  tag = ggplot2::waiver(),
  alt = ggplot2::waiver(),
  fill = ggplot2::waiver(),
  color = ggplot2::waiver(),
  size = ggplot2::waiver(),
  shape = ggplot2::waiver(),
  location = NULL,
  name_col = NULL,
  source = NULL,
  .na = "NA",
  .null = NULL,
  .envir = parent.frame(),
  ...
)

Arguments

title

The text for the title.

subtitle

The text for the subtitle for the plot which will be displayed below the title.

caption

The text for the caption which will be displayed in the bottom-right of the plot by default.

tag

The text for the tag label which will be displayed at the top-left of the plot by default.

alt

Text used for the generation of alt-text for the plot.

fill, color, size, shape

Optional labels for mapped aesthetics. Default to ggplot2::waiver().

location

sf or bbox object or character string

name_col

Column name holding name or identifier for distinct places within the simple feature collection provided to location. Not supported for bbox objects.

source

Data source(s). Not yet used or supported by function.

.na, .null

Additional parameters passed to glue::glue()

.envir

Environment passed to glue::glue(); defaults to parent.frame().

...

A list of new name-value pairs. The name should be an aesthetic.